+++ packages/react-behave/src/components/index.js
@@ -1,6 +1,5 @@
 export { default as ClickOutside } from './ClickOutside';
 export { default as Dropdown } from './Dropdown';
 export { default as MergeRefs } from './MergeRefs';
-export { default as ResizeObserver } from './ResizeObserver';
 export { default as Responsive } from './Responsive';
 export { default as Select } from './Select';