+++ packages/react-behave/src/core/index.js
@@ -1,3 +1,5 @@
 export { default as getDisplayName } from './getDisplayName';
 export { default as getOtherProps } from './getOtherProps';
+export { default as getScreenSize } from './getScreenSize';
 export { default as toInnerRef } from './toInnerRef';
+export { default as watchResize } from './watchResize';